requirements.txt
google-cloud-aiplatform
google-cloud-secret-manager
requests==2.31.0
streamlit==1.31.0